# internal/docs/openapi.yaml
openapi: 3.0.3
info:
  title: go-tari-netmap API
  description: >-
    Tari network peer-topology + node-health mapping API. Polls Tari nodes, walks the
    peer graph via seed-node discovery, and surfaces topology + node-health data for
    mining-pool operators picking peers. Every route documented here is served by
    `internal/api.NewRouter`, mounted at `/api` (via `http.StripPrefix("/api", ...)`
    in `cmd/netmap/main.go`) on the same listener as the HTML dashboard
    (`internal/web`).
  version: "1.0.0"
  contact:
    name: go-tari-netmap
    url: https://github.com/Snipa22/go-tari-netmap
servers:
  - url: https://netmap.supportxtm.com
    description: Production deployment (Caddy-proxied)
  - url: http://localhost:8080
    description: Local development (cmd/netmap's default -addr)
...
paths:
...
  /api/nodes:
    get:
      operationId: listNodes
      tags: [nodes]
      summary: Paginated list of known nodes (privacy-scrubbed)
      description: >-
        Returns a page of `PublicNode`s plus pagination metadata. Every node passes
        through `ScrubNode` — real addresses are only included for nodes whose owner
        opted in via the public submission flow (`discovery_source` `registry_submitted`
        or `both`); a purely `p2p_discovered` node never exposes its real address.
      parameters:
        - name: discovery_source
          in: query
          required: false
          description: Restrict to nodes with this exact discovery_source value.
          schema:
            type: string
            enum: [p2p_discovered, registry_submitted, both]
        - name: owner
          in: query
          required: false
          description: Restrict to nodes where tags.owner exactly equals this value.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: >-
            Page size. Defaults to 100 when omitted or non-positive; capped at 500
            regardless of what's requested.
          schema:
            type: integer
            minimum: 1
            default: 100
            maximum: 500
        - name: offset
          in: query
          required: false
          description: Rows to skip. Defaults to 0.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A page of nodes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListNodesResponse"
        "400":
          description: Malformed `offset` query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Store failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createNode
      tags: [nodes]
      summary: Publicly submit a node for registration (queued for human review)
      description: >-
        Queues a `PendingSubmission` for admin review (see `POST
        /admin/submissions/{id}/approve`) rather than writing directly to the node
        registry. Rate-limited per source IP (`Retry-After` header on 429), with an
        escalating lockout for sources with repeated invalid submissions. Rejects
        private/reserved-IP hosts (basic SSRF hardening) and addresses that are
        already publicly opted-in (409 — contact the operator instead of resubmitting
        someone else's node). A best-effort, non-blocking connectivity probe is kicked
        off for the reviewer's benefit; it never delays this response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNodeRequest"
      responses:
        "202":
          description: Submission accepted and queued for review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingSubmission"
        "400":
          description: >-
            Invalid JSON body, missing host, out-of-range port, or a host that fails
            SSRF validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: This address is already publicly registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Rate-limited or temporarily locked out after repeated invalid submissions.
          headers:
            Retry-After:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: The pending-submission review queue is full (capped at 100 unreviewed rows).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Store failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >-
        HTTP Basic Auth against NETMAP_ADMIN_USER/NETMAP_ADMIN_PASSWORD (see
        internal/adminauth.Wrap). Every /admin/* route fails closed with 503 if these
        env vars aren't both configured — a request never reaches the 401 check at
        all in that case.
    collectorKeyAuth:
      type: apiKey
      in: header
      name: X-Collector-Key
      description: >-
        A per-collector API key checked against an operator-configured map of
        collector name -> key (NETMAP_COLLECTOR_KEYS). Fails closed with 503 if no
        collectors are configured at all.
...
  schemas:
    ApiError:
      type: object
      description: >-
        Shared error-response shape for every route's 400/404/409/429/500/503
        responses — this API returns valid JSON even on error, never a plain-text
        body (except for adminauth/collector-auth's 401/503 gate responses, which are
        plain text via http.Error, upstream of any handler in this file).
      required: [error]
      properties:
        error:
          type: string
          example: node not found
    PublicNode:
      type: object
      description: >-
        The privacy-scrubbed view of a node (see internal/api/privacy.go's ScrubNode).
        Real address strings are NEVER included for a node discovered purely by
        walking the peer graph (discovery_source p2p_discovered) — addresses is
        omitted in that case. Node identity is public_key, not addresses — a
        nil/empty public_key means the node is an unconfirmed placeholder, not an
        error.
      required: [id, discovery_source, tags, first_seen, last_seen, has_ipv4, has_ipv6, has_onion]
      properties:
        id:
          type: string
          format: uuid
        public_key:
          type: string
          format: byte
          description: Base64 of the raw pubkey bytes (Go's default []byte JSON marshaling). Omitted for an unconfirmed node.
        discovery_source:
          $ref: "#/components/schemas/DiscoverySource"
        tags:
          type: object
          additionalProperties: true
        label:
          type: string
          nullable: true
        first_seen:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        has_ipv4:
          type: boolean
        has_ipv6:
          type: boolean
        has_onion:
          type: boolean
        addresses:
          type: array
          items:
            type: string
          description: Only present when discovery_source is registry_submitted or both.
        wallet_http_port:
          type: integer
          nullable: true
          description: >-
            Only ever present under the exact same opt-in gate as addresses above —
            never populated for a node that hasn't opted into public address listing.
...
    DiscoverySource:
      type: string
      enum: [p2p_discovered, registry_submitted, both]
      description: >-
        How a node entered the topology: discovered by walking the peer graph,
        submitted via the public registry API, or both.
    ListNodesResponse:
      type: object
      required: [nodes, total, limit, offset, has_more]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/PublicNode"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
    CreateNodeRequest:
      type: object
      description: >-
        `port` accepts either a JSON number or a numeric JSON string (the dashboard's
        htmx form submits all fields, including `<input type="number">`, as JSON
        strings via the json-enc extension).
      required: [host, port]
      properties:
        host:
          type: string
          description: A public IP or .onion address. Private/reserved IPs are rejected.
        port:
          oneOf:
            - type: integer
            - type: string
          minimum: 1
          maximum: 65535
        label:
          type: string
        owner_tag:
          type: string
    PendingSubmission:
      type: object
      description: One row of the public node-submission review queue.
      required: [id, address, status, submitted_at]
      properties:
        id:
          type: string
          format: uuid
        address:
          type: string
        label:
          type: string
          nullable: true
        owner_tag:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/SubmissionStatus"
        submitted_at:
          type: string
          format: date-time
        reviewed_at:
          type: string
          format: date-time
          nullable: true
        rejection_reason:
          type: string
          nullable: true
        promoted_node_id:
          type: string
          format: uuid
          nullable: true
        probe_attempted_at:
          type: string
          format: date-time
          nullable: true
          description: Nil until the best-effort, informational, async connectivity probe finishes.
        probe_reachable:
          type: boolean
          nullable: true
    SubmissionStatus:
      type: string
      enum: [pending, approved, rejected]
...